+++ src/DataFilters/Grammar/Syntax/StartsWithExpression.cs
@@ -105,5 +105,14 @@
 
         ///<inheritdoc/>
         public override string EscapedParseableString => _lazyEscapedParseableString.Value;
+
+        /// <summary>
+        /// Combines the specified <paramref name="left"/> and <paramref name="right"/> <see cref="StartsWithExpression"/>s into a new <see cref="AndExpression"/>.
+        /// </summary>
+        /// <param name="left"></param>
+        /// <param name="right"></param>
+        /// <returns>a <see cref="AndExpression"/> whose <see cref="BinaryFilterExpression.Left"/> is <paramref name="left"/> and <see cref="BinaryFilterExpression.Right"/> is 
+        /// <paramref name="right"/></returns>
+        public static AndExpression operator +(StartsWithExpression left, EndsWithExpression right) => new (left, right);
     }
 }